extract `useTasksTable`
